#660 - store - split tax module
