+++ core/store/modules/tax/index.js
@@ -1,0 +1,11 @@
+import actions from './actions'
+import mutations from './mutations'
+
+export default {
+  namespaced: true,
+  state: {
+    rules: []
+  },
+  actions,
+  mutations
+}
